refactor: resolve all circular dependencies
