+++ src/configs/results.ts
@@ -1,0 +1,6 @@
+export const results = {
+  passed: 0,
+  failed: 0,
+  skipped: 0,
+  todo: 0,
+};
