+++ src/modules/essentials/poku.ts
@@ -3,7 +3,8 @@ import type { PokuPlugin } from '../../@types/plugin.js';
 import type { Configs, PluginContext, Poku } from '../../@types/poku.js';
 import { join } from 'node:path';
 import { env, hrtime, stdout } from 'node:process';
-import { GLOBAL, results, timespan } from '../../configs/poku.js';
+import { GLOBAL, timespan } from '../../configs/poku.js';
+import { results } from '../../configs/results.js';
 import { reporter } from '../../services/reporter.js';
 import { runTests } from '../../services/run-tests.js';
 import { exit } from '../helpers/exit.js';
