+++ src/parsers/escape-regexp.ts
@@ -1,0 +1,4 @@
+const SAFE_REGEXP = /[.*{}[\]\\]/g;
+
+export const escapeRegExp = (value: string): string =>
+  value.replace(SAFE_REGEXP, '\\$&');
