+++ src/services/run-test-file.ts
@@ -99,6 +99,7 @@ export const runTestFile = async (path: string): Promise<boolean> => {
         const parsedOutputs = parserOutput({
           output,
           result,
+          debug: GLOBAL.configs.debug,
         })?.join('\n');
 
         const total = end[0] * 1e3 + end[1] / 1e6;
