+++ test/integration/no-isolate.test.ts
@@ -1,6 +1,7 @@
 import { join } from 'node:path';
 import { isBuild } from '../__utils__/capture-cli.test.js';
-import { GLOBAL, results } from '../../src/configs/poku.js';
+import { GLOBAL } from '../../src/configs/poku.js';
+import { results } from '../../src/configs/results.js';
 import { assert } from '../../src/modules/essentials/assert.js';
 import { describe } from '../../src/modules/helpers/describe.js';
 import { it } from '../../src/modules/helpers/it/core.js';
